Improve LNURLP status checks
Some LNURLP implementations (such as the LNBits 0.12.x on demo.lnpiggy.com)
don't send a "status" field in their initial response, nor in their invoice reponse.

This was making getInvoiceFromLNAddr() fail,
even though it works find with all other tested clients.

Instead of failing if the status is not present or if it's not OK,
this commit only fails if the status is not OK,
but continues if there is no explicit status field set.